+++ lib/Filter.ts
@@ -1,4 +1,5 @@
 import { DataType } from './DataGridVue'
+import Formatter from './Formatter'
 
 /**
  * @group Filter
@@ -92,6 +93,42 @@ export interface Filter {
   or: FilterCondition[]
   /** Optional {@link Filter} to and with the current one. */
   and: Filter | undefined
+}
+
+/**
+ * @group Filter
+ * @description Compiles a {@link FilterCondition} into a string summary.
+ * @param condition The {@link FilterCondition} to compile.
+ */
+export function CompileFilterConditionSummary(condition: FilterCondition | undefined): string {
+  if (!condition) {
+    return ''
+  }
+  return `${Formatter.fromCamelCase(condition.fieldName)} ${Formatter.fromCamelCase(FilterOperator[condition.operator]).toLocaleLowerCase()} ${
+    condition.value
+  }`
+}
+
+/**
+ * @group Filter
+ * @description Compiles a {@link Filter} into a string summary.
+ * @param filter The {@link Filter} to compile.
+ */
+export function CompileFilterSummary(filter: Filter | undefined): string {
+  let summary = ''
+  if (!filter) {
+    return summary
+  }
+
+  if (filter.or.length) {
+    summary = filter.or.map(CompileFilterConditionSummary).join(' or ')
+  }
+
+  if (filter.and) {
+    summary = `(${CompileFilterSummary(filter.and)}) and ${summary}`
+  }
+
+  return summary
 }
 
 /**