+++ vuepress/shared/changelog.md
@@ -14,25 +14,34 @@
     * `StorageService.getGridState` -> `StorageService.getGridStateAsync`
     * `StorageService.setGridState` -> `StorageService.setGridStateAsync`
   :::
-  ::: tip New Features
+  ::: tip New Feature Highlights
   * New footer slots
     * `footer` - entire footer
     * `footer-page-size-select` - footer page size select
     * `footer-additional-content` - additional content between page size select and total item text
     * `footer-total-items` - total items text
+  * [#4](https://github.com/nruffing/data-grid-vue/issues/4) accessibility improvements
+    * Add missing aria-label attributes
+    * Allow header to be navigated via keyboard
+    * Allow sort and reorder actions to be performed via keyboard when header cell has focus
+      * Space or enter to cycle through sort options
+      * Left and right arrow to reorder
+    * Header options can be triggered via space or enter
   :::
   ::: details More changes
-  * Update [`debounce`](https://www.npmjs.com/package/debounce/v/2.0.0) dependency to new major version v2.0.0 which requires node v18 ([current oldest supported version](https://nodejs.org/en/about/previous-releases)).
-  * Update ['dragon-drop-vue'](https://www.npmjs.com/package/dragon-drop-vue) dependency to v0.2.0
+  * Add additional parameters to the `options-header`, `options-header-filter-options-shown`, and `options-header-clear-filters` slots.
   * `ServerSideStorageService` now has a generic type constraint to allow any type to be used for the user identifier sent in the request to get and set grid state.
   * `ServerSideStorageService.getGridState` and `ServerSideDataService.getPage` will now only try to deserialize the response body as JSON if the status code is `200 OK` and the response `Content-Type` header is `application/json`.
+  * Dynamic column header titles
+  * Entire header cell is now the click target for a sort.
   * Default color values of CSS variables are now defined in hex.
   * Default accent color slightly altered to match documentation site.
   * Improved default layout styles of custom column filters set using the `filter-{fieldName}` slot.
-  * [#4](https://github.com/nruffing/data-grid-vue/issues/4) accessibility improvements
   * Documentation site now includes documentation for [DataGridVueDotnet](https://github.com/nruffing/data-grid-vue-dotnet).
   * Repo now uses [`pnpm`](https://pnpm.io/) where the `data-grid-vue` package is build from the root workspace and the documentation site and dev app are nested workspaces.
   * Now being built with `vite` [v5](https://vitejs.dev/blog/announcing-vite5).
+  * Update [`debounce`](https://www.npmjs.com/package/debounce/v/2.0.0) dependency to new major version v2.0.0 which requires node v18 ([current oldest supported version](https://nodejs.org/en/about/previous-releases)).
+  * Update ['dragon-drop-vue'](https://www.npmjs.com/package/dragon-drop-vue) dependency to v0.2.0
   * Documentation site theme update
   * Documentation site is now setup up as a [progressive web app](https://developer.mozilla.org/en-US/docs/Web/Progressive_web_apps).
   :::